Add files via upload
Task:
First link this dataset
 
https://www.kaggle.com/rounakbanik/the-movies-dataset#credits.csv
 
which has the actor/actresses from each film 
 
with 
 
https://www.kaggle.com/rounakbanik/the-movies-dataset#movies_metadata.csv
 
which gives you the film's year of release. 
 
Match based on the strings of actor names with the strings of actor names in name.basics.tsv here
 
https://datasets.imdbws.com/
 
This will allow you to match actors with their birthyears.
 
 
The ultimate goal is to record the age of all actors appearing in a film in a given year.